+++ src/imio/dashboard/interfaces.py
@@ -22,15 +22,3 @@ class ICustomViewFieldsVocabulary(Interface):
         """
           Get every IFacetedColumn z3c.table columns.
         """
-
-
-class NotDashboardContextException(Exception):
-    """ To be raised when a context has no faceted view defined on it. """
-
-
-class NoFacetedViewDefinedException(NotDashboardContextException):
-    """ To be raised when a context has no faceted view defined on it. """
-
-
-class NoCollectionWidgetDefinedException(NotDashboardContextException):
-    """ To be raised when a context has no collection widget defined on it. """